Permissions update (huge!)

+++ app/Http/Repos/UserRepo.php
@@ -6,13 +6,15 @@ use App\User;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Str;
 use Illuminate\Support\Facades\DB;
+use Spatie\Permission\Models\Role;
+use Spatie\Permission\Models\Permission;
 
 class UserRepo
 {
-    public function AddUserToAccountUser($contact_id, $user_id) {
-        $accountUser = \App\AccountUser::where('contact_id', $contact_id)->first();
+    public function AddUserToAccountUser($contactId, $userId) {
+        $accountUser = \App\AccountUser::where('contact_id', $contactId)->first();
 
-        $accountUser->user_id = $user_id;
+        $accountUser->user_id = $userId;
 
         $accountUser->save();
     }
@@ -20,47 +22,57 @@ class UserRepo
     public function ChangePassword($userId, $password) {
         $old = $this->GetById($userId);
 
-        $old->password = $password;
+        $old->password = Hash::make($password);
         $old->save();
     }
 
-    public function CountAccountUsers($account_id) {
-        $count = \App\AccountUser::where('account_id', $account_id)->count();
+    public function CountAccountUsers($accountId) {
+        $count = \App\AccountUser::where('account_id', $accountId)->count();
 
         return $count;
     }
 
-    public function DeleteAccountUserByContactId($contactId) {
-        $accountUser = AccountUser::where('contact_id', $contactId)->first();
-        $user = User::where('user_id', $accountUser->user_id)->first();
+    public function DeleteAccountUser($contactId, $accountId) {
+        $userId = AccountUser::where('contact_id', $contactId)
+            ->where('account_id', $accountId)
+            ->first()->userId;
 
-        $user->delete();
-        $accountUser->delete();
+        /**
+         * Secondary full request for delete is required, otherwise it will delete all entries matching the primary key ('contact_id')
+         */
+        $accountUser = AccountUser::where('contact_id', $contactId)
+            ->where('account_id', $accountId)
+            ->delete();
 
-        $contactRepo = new ContactRepo();
-        $contactRepo->Delete($contactId);
+        if(AccountUser::where('contact_id', $contactId)->count() == 1) {
+            $user = User::where('user_id', $userId)->first();
+            $user->delete();
+            $contactRepo = new ContactRepo();
+            $contactRepo->Delete($contactId);
+        }
     }
 
-    public function GetAccountUserByContactId($contact_id) {
-        $accountUser = \App\AccountUser::where('contact_id', $contact_id);
+    public function GetAccountUser($contactId, $accountId) {
+        $accountUser = AccountUser::where('contact_id', $contactId)
+            ->where('account_id', $accountId);
 
         return $accountUser->first();
     }
 
-    public function GetAccountUserIds ($account_id) {
-        $query = \App\AccountUser::where('account_id', $account_id)
-            ->leftJoin('email_addresses', 'email_addresses.contact_id', '=', 'account_users.contact_id')
-            ->leftJoin('phone_numbers', 'phone_numbers.contact_id', '=', 'account_users.contact_id');
-        $accountUsers['contact_ids'] = $query->pluck('account_users.contact_id');
-        $accountUsers['user_ids'] = $query->pluck('user_id');
-        $accountUsers['email_ids'] = $query->pluck('email_address_id');
-        $accountUsers['phone_ids'] = $query->pluck('phone_number_id');
+    // public function GetAccountUserIds ($account_id) {
+    //     $query = \App\AccountUser::where('account_id', $account_id)
+    //         ->leftJoin('email_addresses', 'email_addresses.contact_id', '=', 'account_users.contact_id')
+    //         ->leftJoin('phone_numbers', 'phone_numbers.contact_id', '=', 'account_users.contact_id');
+    //     $accountUsers['contact_ids'] = $query->pluck('account_users.contact_id');
+    //     $accountUsers['user_ids'] = $query->pluck('user_id');
+    //     $accountUsers['email_ids'] = $query->pluck('email_address_id');
+    //     $accountUsers['phone_ids'] = $query->pluck('phone_number_id');
 
-        return $accountUsers;
-    }
+    //     return $accountUsers;
+    // }
 
-    public function GetAccountUsers($account_id) {
-        $accountUsers = AccountUser::where('account_id', '=', $account_id)
+    public function GetAccountUsers($accountId) {
+        $accountUsers = AccountUser::where('account_id', '=', $accountId)
             ->leftJoin('contacts', 'account_users.contact_id', '=', 'contacts.contact_id')
             ->leftJoin('users', 'account_users.user_id', '=', 'users.user_id')
             ->leftJoin('email_addresses', 'account_users.contact_id', '=', 'email_addresses.contact_id')
@@ -74,14 +86,16 @@ class UserRepo
                 'email_addresses.email as primary_email',
                 'phone_numbers.phone_number as primary_phone',
                 'contacts.position as position',
-                'account_users.is_primary as is_primary'
+                'account_users.is_primary as is_primary',
+                'users.is_enabled as enabled',
+                DB::raw('(select count(*) from account_users where account_users.contact_id = contacts.contact_id) as belongs_to_count')
             );
 
         return $accountUsers->get();
     }
 
-    public function GetById($id) {
-        $user = User::where('user_id', '=', $id)->first();
+    public function GetById($userId) {
+        $user = User::where('user_id', '=', $userId)->first();
 
         return $user;
     }
@@ -93,12 +107,17 @@ class UserRepo
         return $user;
     }
 
+    public function GetUserByPrimaryEmail($emailAddress) {
+        $user = User::where('email', $emailAddress);
+
+        return $user->first();
+    }
+
     public function Insert($user) {
         $new = new User;
 
         $user = array_merge($user, array(
             'password' => Hash::make(Str::random(15)),
-            'is_locked' => false,
             'login_attempts' => 0,
             'remember_token' => null
         ));
@@ -108,36 +127,36 @@ class UserRepo
         return $new;
     }
 
-    public function InsertAccountUser($account_user) {
+    public function InsertAccountUser($accountUser) {
         $new = new \App\AccountUser;
-        
-        $new = $new->create($account_user);
+
+        $accountUserCount = AccountUser::where('account_id', $accountUser['account_id'])->count();
+        if($accountUserCount > 0)
+            $accountUser['is_primary'] = 0;
+
+        $new = $new->create($accountUser);
 
         return $new;
     }
 
-    public function SetPrimaryAccountUser($contact_id) {
-        $newPrimary = \App\AccountUser::where($contact_id, 'contact_id')->first();
-        $oldPrimary = \App\AccountUser::where('account_id', $newPrimary->account_id)
-            ->where('is_primary', true)
-            ->first();
-        if(isset($oldPrimary->contact_id)) {
-            $oldPrimary->is_primary = false;
-            $oldPrimary->save();
-        }
-        $newPrimary->is_primary = true;
-        $newPrimary->save();
+    public function LinkAccountUser($contactId, $accountId) {
+        $userId = AccountUser::where('contact_id', $contactId)->pluck('user_id');
 
-        return $newPrimary;
+        $accountUserCount = AccountUser::where('account_id', $accountId)->count();
+        $new = new \App\AccountUser;
+
+        $new->create(['account_id' => $accountId, 'contact_id' => $contactId, 'is_primary' => $accountUserCount > 0, 'user_id' => $userId[0]]);
+
+        return $new;
     }
 
-    public function Update($user) {
+    public function Update($user, $updatePermissions = false) {
         $old = $this->GetById($user['user_id']);
 
         // no support for usernames atm
         // $old->username = $user['username'];
-        if(isset($user->is_locked))
-            $old->is_locked = $user['is_locked'];
+        if($updatePermissions)
+            $old->is_enabled = $user['is_enabled'];
         $old->email = $user['email'];
 
         $old->save();
